style : HomeScreen HomeScreen 스타일 분리
