+++ sospj/screens/HomeScreen.js
@@ -1,7 +1,10 @@
 import React from 'react';
-import { View, Text, Button, StyleSheet, ScrollView, Linking, Alert, Pressable } from 'react-native';
+import { View, Text, Button, ScrollView, Linking, Alert, Pressable, TouchableOpacity } from 'react-native';
+import styles from '../teststyle/HomeStyle';
 import NfcScanner from '../NFC/nfcScanner';
-import { TouchableOpacity } from 'react-native-gesture-handler';
+
+// Your component code remains the same, just remove the styles object and its declaration.
+
 
 function callNumber(phoneNumber) {
   // 전화번호에서 하이픈('-') 제거
@@ -81,67 +84,6 @@ function HomeScreen({ navigation }) {
   );
 }
 
-const styles = StyleSheet.create({
-  header: {
-    fontSize: 22,
-    margin:10
-  },
-  scrollView: {
-    width: '100%',
-    marginBottom: 20,
-  },
-  contactItem: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: 10,
-    borderBottomWidth: 1,
-    borderBottomColor: '#ccc',
-  },
-  contactText: {
-    fontSize: 18,
-  },
-  row: {
-    flexDirection: 'row',
-    marginBottom: 5,
-  },
-  row1: {
-    flexDirection: 'row',
-    marginBottom: 5,
-    justifyContent: 'space-between',
-    width:'100%'
-  },
-  label: {
-    fontWeight: 'bold',
-  },
-  value: {
-    marginLeft: 5,
-  },
-  button: {
-    padding: 10,
-    borderRadius: 5,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  text: {
-    color: '#ffffff',
-    fontSize: 16,
-  },
-  emergencyButtonsContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    marginTop: 20,
-  },
-  emergencyButton: {
-    padding: 10,
-    borderRadius: 5,
-    margin:20,
-    alignItems: 'center',
-  },
-  emergencyButtonText: {
-    color: '#ffffff',
-    fontWeight: 'bold',
-  },
-});
+
 
 export default HomeScreen;
